added additional info
