+++ sfucompass/src/controllers/lib/middleware.js
@@ -1,16 +1,17 @@
 import nextConnect from 'next-connect'
-import passport from '../lib/passport'
-import session from '../lib/session'
+import passport from './passport'
+import session from './session'
 
+// middleware to route passport.js to next.js
 const auth = nextConnect()
   .use(
     session({
-      name: 'sess',
+      name: 'session',
       secret: process.env.local.AUTH0_SECRET,
       cookie: {
         maxAge: 60 * 60 * 8, // 8 hours,
         httpOnly: true,
-        secure: process.env.NODE_ENV === 'production',
+        secure: process.env.NODE_ENV != 'production',
         path: '/',
         sameSite: 'lax',
       },
